增加所有区域

+++ card.h
@@ -162,7 +162,7 @@ public:
 	uint32 sum_param;
 	uint32 position_param;
 	uint32 spsummon_param;
-	uint32 to_field_param;
+	uint64 to_field_param;
 	uint8 attack_announce_count;
 	uint8 direct_attackable;
 	uint8 announce_count;
